Add packages tasks and dukeexceptions
