+++ src/main/java/dukeexceptions/DukeException.java
@@ -1,3 +1,5 @@
+package dukeexceptions;
+
 public class DukeException extends Exception {
 
     public DukeException(String message) {
