+++ src/main/java/dukeexceptions/NoDescriptionException.java
@@ -1,3 +1,5 @@
+package dukeexceptions;
+
 public class NoDescriptionException extends DukeException {
     public NoDescriptionException(String temp) {
         super(String.format(":-( OOPS! The description of %s cannot be empty. Try again!", temp));
