+++ src/main/java/dukeexceptions/NoSuchCommandException.java
@@ -1,3 +1,5 @@
+package dukeexceptions;
+
 public class NoSuchCommandException extends DukeException {
     public NoSuchCommandException() {
         super(":-( OOPS! No such command exists. Try again!");
