+++ src/main/java/tasks/Events.java
@@ -1,3 +1,5 @@
+package tasks;
+
 public class Events extends Task {
     protected String description;
     protected boolean isDone;
