+++ src/main/java/tasks/Todos.java
@@ -1,3 +1,5 @@
+package tasks;
+
 public class Todos extends Task {
     protected String description;
     protected boolean isDone;
